python-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
httpx==0.27.0
python-dotenv==1.0.1
openai==1.40.0
finnhub-python==2.4.20
sse-starlette==2.1.0
pytest==8.2.0
pytest-asyncio==0.23.8
